+++ relay-event-normalization/src/event.rs
@@ -1372,9 +1372,10 @@ fn normalize_force_trace_context(event: &mut Event) {
     let contexts = event.contexts.get_or_insert_with(Contexts::new);
     let trace = contexts.get_or_default::<TraceContext>();
 
-    let trace_id = trace
-        .trace_id
-        .get_or_insert_with(|| TraceId::from(*event.id.get_or_insert_with(Default::default)));
+    let trace_id = trace.trace_id.get_or_insert_with(|| {
+        TraceId::try_from(*event.id.get_or_insert_with(Default::default))
+            .unwrap_or_else(|_| TraceId::random())
+    });
     let _ = trace
         .span_id
         .get_or_insert_with(|| SpanId::derive_from_trace_id(trace_id));
@@ -1423,8 +1424,8 @@ fn normalize_mobile_measurements(measurements: &mut Measurements) {
 }
 
 const APP_START_SOURCES: [(&str, Option<&str>); 5] = [
-    ("app_start_cold", Some("cold")),
-    ("app_start_warm", Some("warm")),
+    (APP_START_COLD, Some("cold")),
+    (APP_START_WARM, Some("warm")),
     (APP__VITALS__START__VALUE, None),
     (APP__VITALS__START__COLD__VALUE, None),
     (APP__VITALS__START__WARM__VALUE, None),